Put each CLI command in its own file
