+++ lib/commands/build.js
@@ -1,0 +1,3 @@
+module.exports = function(args, options) {
+  this.watch(args, options, true)
+}
