+++ lib/commands/docs.js
@@ -1,0 +1,5 @@
+var open = require('open');
+
+module.exports = function() {
+  open('http://foundation.zurb.com/apps/docs');
+}
