+++ lib/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  help: require('./commands/help'),
+  new: require('./commands/new'),
+  update: require('./commands/update'),
+  watch: require('./commands/watch'),
+  build: require('./commands/build'),
+  docs: require('./commands/docs')
+}
